+++ stockMate/src/main/webapp/WEB-INF/views/location.jsp
